# Conflicts:
#	app/Http/Controllers/API/TelegramApiController.php
#	app/Services/Telegram/Client.php
#	routes/api/v1.php
#	routes/web.php

+++ database/migrations/2026_01_05_000001_create_news_groups_table.php
@@ -9,7 +9,7 @@ return new class extends Migration {
     {
         Schema::create('news_groups', function (Blueprint $table): void {
             $table->id();
-            $table->tinyInteger('language')->index(); // Language::RU
+            $table->smallInteger('language')->index(); // Language::RU
             $table->boolean('active')->default(false);
             $table->string('title');
             $table->string('code')->index();
